[SPARK-48112][CONNECT] Expose session in SparkConnectPlanner to plugins
### What changes were proposed in this pull request?

This PR makes `session` in `SparkConnectPlanner` public so that it can be accessed by Spark Connect plugins. This method was public before but was made private. To prevent this from happening I have added `DeveloperApi` as an annotation.

### Why are the changes needed?

To allow the development of Spark Connect plugins.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

N/A

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #46363 from tomvanbussel/SPARK-48112.

Authored-by: Tom van Bussel <tom.vanbussel@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>