[SPARK-48176][SQL] Adjust name of FIELD_ALREADY_EXISTS error condition
### What changes were proposed in this pull request?

Rename `FIELDS_ALREADY_EXISTS` to `FIELD_ALREADY_EXISTS`.

### Why are the changes needed?

Though it's not meant to be a proper English sentence, `FIELDS_ALREADY_EXISTS` is grammatically incorrect. It should either be "fields already exist[]" or "field[] already exists". I opted for the latter.

### Does this PR introduce _any_ user-facing change?

Yes, it changes the name of an error condition.

### How was this patch tested?

CI only.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #46510 from nchammas/SPARK-48176-field-exists-error.

Authored-by: Nicholas Chammas <nicholas.chammas@gmail.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>