[SPARK-48239][INFRA] Update the release docker image to follow what we use in Github Action jobs
### What changes were proposed in this pull request?

We have Github Action jobs to test package building and doc generation, but the execution environment is different from what we use for the release process.

This PR updates the release docker image to follow what we use in Github Action: https://github.com/apache/spark/blob/master/dev/infra/Dockerfile

Note: it's not exactly the same, as I have to do some modification to make it usable for the release process. In the future we should have a better way to unify these two docker files.

### Why are the changes needed?

to make us be able to release

### Does this PR introduce _any_ user-facing change?

no

### How was this patch tested?

manually

### Was this patch authored or co-authored using generative AI tooling?

no

Closes #46534 from cloud-fan/re.

Authored-by: Wenchen Fan <wenchen@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>