[SPARK-48031][SQL] Support view schema evolution
### What changes were proposed in this pull request?

Add the following syntax to CREATE VIEW... WITH SCHEMA ...:
```
CREATE [ OR REPLACE ] [ TEMPORARY ] VIEW [ IF NOT EXISTS ] view_name
    [ column_list ]
    [ schema_binding ]
    [ COMMENT view_comment ]
    [ TBLPROPERTIES clause ]
    AS query

schema_binding
   WITH SCHEMA { BINDING | [TYPE] EVOLUTION | COMPENSATION }

column_list
   ( { column_alias [ COMMENT column_comment ] } [, ...] )
```

Allow changing of the schema binding in:
```
ALTER VIEW view_name  schema_binding
```

The semantic is:

Schema_binding
Optionally specifies how the view adapts to changes to the schema of the query due to changes in the underlying object definitions.
This clause is not supported for temporary views.

- BINDING
The view will become invalid if the query column-list changes except for the following conditions:
The column-list includes a star clause, and there are additional columns. These additional columns are ignored.
The type of one or more columns changed in a way that allows them to be safely cast to the types using implicit casting rules.

- COMPENSATION
The view will become invalid if the query column list changes except for the following conditions:
The column-list includes a star clause, and there are additional columns. These additional columns are ignored.
The type of one or more columns changed in a way that allows them to be cast using explicit cast rules.
This is the default behavior.

- TYPE EVOLUTION
The view will adopt any changes to types in the query column list into is own definition when such a change is detected upon reference of the view.

- EVOLUTION
Behaves like TYPE EVOLUTION and also adopts changes in column names or added and dropped columns if the view does not include an explicit column list.
The view will only be invalidated if the query cannot be parsed anymore, or the optional view column_list does not match the number of expressions in the query select-list anymore.

We also introduce a new SQL Config:
spark.sql.defaultViewSchemaBinding
 Control the default behavior of views when the underlying schema changes.
 Valid values are:
* COMPENSATION - Any supported casts.
* DISABLED - Disable the feature.

### Why are the changes needed?

Schema changes are a frequent occurrence, especially when ingesting data.
In the course of it most frequently:

- Types may need to be widened
- Columns are added to tables
- Fields are added to structs

The traditional SCHEMA BINDING behavior makes schema evolution very hard since it invalidates views agressively.
This causes erros, requiring user intervention.
Allowing views to be created to "roll with the punches" or tolerate changes in the underlying schema improves uptime.

### Does this PR introduce _any_ user-facing change?

Yes, this is a new feature with new grammar and a new config to influence its default.

### How was this patch tested?

New tests are added

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #46267 from srielau/SPARK-48031-view-evolution.

Authored-by: Serge Rielau <serge@rielau.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>