[SPARK-41794][SQL] Add `try_remainder` function and re-enable column tests
### What changes were proposed in this pull request?
As part of re-enabling the ANSI mode tests for Spark Connect, we discovered that we don't have an equivalent for `try_*` for the remainder of operations. This patch adds the `try_remainder` function in Scala, Python, and Spark Connect and adds the required testing.

### Why are the changes needed?
ANSI and Spark 4

### Does this PR introduce _any_ user-facing change?
Yes, it adds the `try_remainder` function that behaves according to ANSI for division by zero.

### How was this patch tested?
Added new UT and E2E tests.

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #46434 from grundprinzip/grundprinzip/SPARK-41794.

Lead-authored-by: Martin Grund <martin.grund@databricks.com>
Co-authored-by: Martin Grund <grundprinzip@gmail.com>
Co-authored-by: Hyukjin Kwon <gurwls223@apache.org>
Signed-off-by: Gengliang Wang <gengliang@apache.org>