[SPARK-48214][INFRA] Ban import `org.slf4j.Logger` & `org.slf4j.LoggerFactory`
### What changes were proposed in this pull request?
The pr aims to ban import `org.slf4j.Logger` & `org.slf4j.LoggerFactory`.

### Why are the changes needed?
After the migration of structured logs on the `java side` is completed, we need to ban import `org.slf4j.Logger` & `org.slf4j.LoggerFactory` in the code to avoid the log format that is not written as required in the future new java code.

### Does this PR introduce _any_ user-facing change?
Yes, only for spark developers.

### How was this patch tested?
- Manually test.
- Pass GA.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #46502 from panbingkun/ban_import_slf4j.

Authored-by: panbingkun <panbingkun@baidu.com>
Signed-off-by: Gengliang Wang <gengliang@apache.org>