[SPARK-45009][SQL][FOLLOW UP] Add error class and tests for decorrelation of predicate subqueries in join condition which reference both join child
### What changes were proposed in this pull request?

This is a follow up PR for https://github.com/apache/spark/pull/42725, which decorrelates predicate subqueries in join conditions. I forgot to add the error class definition for the case where the subquery references both join children, and test cases for it.

### Why are the changes needed?

To show a clear error message when the condition is hit.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Added SQL test and golden files.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #46708 from andylam-db/follow-up-decorrelate-subqueries-in-join-cond.

Authored-by: Andy Lam <andy.lam@databricks.com>
Signed-off-by: Gengliang Wang <gengliang@apache.org>