[SPARK-47361][SQL] Derby: Calculate suitable precision and scale for DECIMAL type
### What changes were proposed in this pull request?

When storing `decimal(p, s)` to derby, if `p > 31`, `s` is wrongly hardcoded to `5` which is the assumed default scale of derby decimal. Actually, 0 is the default scale, 5 is the default precision https://db.apache.org/derby/docs/10.13/ref/rrefsqlj15260.html

This PR calculates a suitable scale to make room for precision.

### Why are the changes needed?

avoid precision loss

### Does this PR introduce _any_ user-facing change?

Yes, but derby is rare in production environments, and the new mapping are compatible for most usecases

### How was this patch tested?

new tests

### Was this patch authored or co-authored using generative AI tooling?
no

Closes #46776 from yaooqinn/SPARK-48439.

Authored-by: Kent Yao <yao@apache.org>
Signed-off-by: yangjie01 <yangjie01@baidu.com>