[SPARK-45891] Describe shredding scheme for Variant
### What changes were proposed in this pull request?

For the Variant data type, we plan to add support for columnar storage formats (e.g. Parquet) to write the data shredded across multiple physical columns, and read only the data required for a given query. This PR merges a document describing the approach we plan to take. We can continue to update it as the implementation progresses.

### Why are the changes needed?

When implemented, can allow much better performance when reading from columnar storage. More detail is given in the document.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

It is internal documentation, no testing should be needed.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #46831 from cashmand/SPARK-45891.

Authored-by: cashmand <david.cashman@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>