[SPARK-48447][SS] Check state store provider class before invoking the constructor
### What changes were proposed in this pull request?
This PR adds a validity check to the StateStoreProvider class before calling its constructor. A new runtime exception was created to report this issue.

### Why are the changes needed?
This is a security improvement to a user-facing API. Users are only allowed to use class that extends `org.apache.spark.sql.execution.streaming.state.StateStoreProvider` as StateStoreProvider.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
This PR comes with a new test to test that invalid class cannot pass the security check. The fact that other tests in StreamingQuerySuite can pass implies that valid class can pass the check.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #46791 from eason-yuchen-liu/check-statestoreprovider.

Authored-by: Yuchen Liu <yuchen.liu@databricks.com>
Signed-off-by: Jungtaek Lim <kabhwan.opensource@gmail.com>