[SPARK-48503][SQL] Fix invalid scalar subqueries with group-by on non-equivalent columns that were incorrectly allowed
### What changes were proposed in this pull request?

Fixes CheckAnalysis to reject invalid scalar subquery group-bys that were previously allowed and returned wrong results.

For example, this query is not legal and should give an error, but instead we incorrectly allowed it and it returns wrong results prior to this PR (full repro with table data in the jira):

```
select *, (select count(*) from y where y1 > x1 group by y1) from x;
```

It returns two rows, even though there's only one row of x. The correct result is an error, because there is more than one row returned by the scalar subquery.

Another problem case is if the correlation condition is an equality but it's under another operator like an OUTER JOIN or UNION. Various other expressions that are not equi-joins between the inner and outer fields hit this too, e.g. `where y1 + y2 = x1 group by y1`. See the comments in the code and the tests for more examples.

This PR fixes the logic which checks for valid vs invalid group-bys. However, note that this new logic may block some queries that are actually valid, for example `a + 1 = outer(b)` is valid but would be rejected. Therefore, we add a conf flag that can be used to restore the legacy behavior, as well as logging for when the legacy behavior is used and differs from the new behavior. (In general, to accurately run valid queries and reject invalid queries, the check must be moved from compile-time to run-time - see https://issues.apache.org/jira/browse/SPARK-48501.)

This is a longstanding bug. The bug is in CheckAnalysis in checkAggregateInScalarSubquery. It allows grouping columns that are present in correlation predicates, but doesn’t check whether those predicates are equalities -  because when that code was written, non-equality correlation wasn’t allowed. Therefore, this bug has existed since non-equality correlation was added (~2 years ago).

### Why are the changes needed?
Fix invalid queries returning wrong results

### Does this PR introduce _any_ user-facing change?
Yes, block subqueries with invalid group-bys.

### How was this patch tested?
Add tests

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #46839 from jchen5/scalar-subq-gby.

Authored-by: Jack Chen <jack.chen@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>