[SPARK-48602][SQL] Make csv generator support different output style with spark.sql.binaryOutputStyle
### What changes were proposed in this pull request?

In SPARK-47911, we introduced a universal BinaryFormatter to make binary output consistent
across all clients, such as beeline, spark-sql, and spark-shell, for both primitive and nested binaries.

But unfortunately, `to_csv` and `csv writer` have interceptors for binary output which is hard-coded to use `SparkStringUtils.getHexString`. In this PR we make it also configurable.

### Why are the changes needed?

feature parity

### Does this PR introduce _any_ user-facing change?

Yes, we have make spark.sql.binaryOutputStyle work for csv but the AS-IS behavior is kept.

### How was this patch tested?

new tests

### Was this patch authored or co-authored using generative AI tooling?
no

Closes #46956 from yaooqinn/SPARK-48602.

Authored-by: Kent Yao <yao@apache.org>
Signed-off-by: Kent Yao <yao@apache.org>