[SPARK-45685][SQL][FOLLOWUP] Add handling for `Stream` where `LazyList.force` is called
### What changes were proposed in this pull request?
Refer to the suggestion of https://github.com/apache/spark/pull/43563#pullrequestreview-2114900378, this pr add handling for Stream where LazyList.force is called

### Why are the changes needed?
Even though `Stream` is deprecated in 2.13, it is not _removed_ and thus is is possible that some parts of Spark / Catalyst (or third-party code) might continue to pass around `Stream` instances. Hence, we should restore the call to `Stream.force` where `.force` is called on `LazyList`, to avoid losing the eager materialization for Streams that happen to flow to these call sites. This is also a guarantee of compatibility.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Add some new tests

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #46970 from LuciferYang/SPARK-45685-FOLLOWUP.

Authored-by: yangjie01 <yangjie01@baidu.com>
Signed-off-by: yangjie01 <yangjie01@baidu.com>