[SPARK-48302][PYTHON] Preserve nulls in map columns in PyArrow Tables
### What changes were proposed in this pull request?
This is a small follow-up to #46529. It fixes a known issue affecting PyArrow Tables passed to `spark.createDataFrame()`. After this PR, if the user is running PyArrow 17.0.0 or higher, null values in MapArray columns containing nested fields or timestamps will be preserved.

### Why are the changes needed?
Before this PR, null values in MapArray columns containing nested fields or timestamps are replaced by empty lists when a PyArrow Table is passed to `spark.createDataFrame()`.

### Does this PR introduce _any_ user-facing change?
It prevents loss of nulls in the case described above. There are no other user-facing changes.

### How was this patch tested?
A test is included.

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #46837 from ianmcook/SPARK-48302.

Authored-by: Ian Cook <ianmcook@gmail.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>